// libs/features/violations/src/lib/shared/enums/document-transaction-type.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum DocumentTransactionType {
  MODIFY_VIOLATION_TYPE = 'CHANGE_VIOLATION',
  CANCEL_VIOLATION_TYPE = 'Cancel Violation',
  CANCEL_TRANSACTION_TYPE = 'CANCEL_VIOLATION',
  MODIFY_TRANSACTION_TYPE = 'MODIFY_VIOLATION',
  CHANGE_ENGAGEMENT_TYPE = 'CHANGE_ENGAGEMENT',
  VIOLATION_REPORT_TYPE = 'VIOLATION_REPORT',
  VIOLATION_LETTER_TYPE = 'VIOLATION_LETTER',
  VIOLATION_DOCUMENT = 'VIOLATION_DOCUMENT',
  REGISTER_VIOLATION_THROUGH_FO = 'REGISTER_VIOLATION_THROUGH_FO',
  REPORT_VIOLATION = 'REPORT_VIOLATION',
  APPEAL_ON_VIOLATION = 'APPEAL_ON_VIOLATION'
}