// libs/features/violations/src/lib/shared/enums/violations-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
export enum ViolationsEnum {
  OTHER_REASON = 'Others',
  ONLY_VC_ONE = 99,
  BOTH_VCS = 69,
  ALL_THREE_VC = 39,
  ROLE_CSR = 'Customer Service Representative',
  VIOLATION_TYPE = 'violationType',
  CONTRIBUTOR_NUMBER = 'contributorNumber',
  BOOLEAN_YES = 'Yes',
  BOOLEAN_NO = 'No',
  RETURN = 'RETURN',
  NO_PENALTY_POPUP_YES = 'YES',
  RASED = 'Rased',
  MANUAL = 'Manual',
  DATE_GREGORIAN = 'Gregorian',
  DATE_HIJIRI = 'Hijiri',
  BENEFIT_TYPE_ALL = 'All',
  ESTABLISHMENT_REGISTERED = 'Registered',
  ESTABLISHMENT_REOPEN = 'Reopened' //1012
}