// libs/features/violations/src/lib/shared/enums/violations-wizard-types.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

/* Form wizard types enum */

export enum ViolationsWizardTypes {
  DOCUMENTS = 'VIOLATIONS.VIOLATION-DOCUMENTS',
  VIOLATION_DETAILS = 'VIOLATIONS.VIOLATION-DETAIL'
}